refactor: Extract common component for displaying photos in different PostWizard steps
